Add a header file with operation struct

+++ ArithmeticCalculations/shuntingYard.h
@@ -5,4 +5,4 @@
 using namespace std;
 string getInput(int argc, char* argv[]);
 string tocken(string& str);
-int calculate(string str);
+double calculate(string str);
